getRecipes() route and data model
